Create cluster.py
This program read data set from blogdata.ods and try to cluster the RSS feed names in first column of data set using the words count. It uses kmean algorithm to do the necessary clustering. The feature vector is multi dimensional, where each dimension is represented by the word in a column of bloddata.ods. The kmean algorithm randomly makes 3 cluster centres and tries to calculate their distance with the feature vectors and moves the cluster centroids according to mean values in each iteration.